incorps timer into watering component
