--- src/components/timer.js
@@ -1,75 +1,0 @@
-import React, { Component } from 'react';
-import Card from '@material-ui/core/Card';
-import CardContent from '@material-ui/core/CardContent';
-import CardHeader from '@material-ui/core/CardHeader';
-import Typography from '@material-ui/core/Typography';
-import Amplify, { API, PubSub } from 'aws-amplify';
-import { AWSIoTProvider } from '@aws-amplify/pubsub/lib/Providers';
-import CircularProgress from '@material-ui/core/CircularProgress';
-import ms from "pretty-ms"
-
-class Timer extends Component {
-
-  constructor(props){
-    super(props);
-    this.state = {
-        time: 0,
-        isOn: false,
-        start: 0
-      }
-      this.startTimer = this.startTimer.bind(this)
-      this.stopTimer = this.stopTimer.bind(this)
-      this.resetTimer = this.resetTimer.bind(this)
-  }
-
-
-  componentDidMount(){
-
-  }
-
-  startTimer() {
-    this.setState({
-      isOn: true,
-      time: this.state.time,
-      start: Date.now() - this.state.time
-    })
-    this.timer = setInterval(() => this.setState({
-      time: Date.now() - this.state.start
-    }), 1);
-  }
-
-  stopTimer() {
-    this.setState({isOn: false})
-    clearInterval(this.timer)
-  }
-
-  resetTimer() {
-    this.setState({time: 0, isOn: false})
-  }
-
-  render() {
-    let start = (this.state.time == 0) ?
-      <button onClick={this.startTimer}>start</button> :
-      null
-    let stop = (this.state.time == 0 || !this.state.isOn) ?
-      null :
-      <button onClick={this.stopTimer}>stop</button>
-    let resume = (this.state.time == 0 || this.state.isOn) ?
-      null :
-      <button onClick={this.startTimer}>resume</button>
-    let reset = (this.state.time == 0 || this.state.isOn) ?
-      null :
-      <button onClick={this.resetTimer}>reset</button>
-    return(
-      <div>
-        <h3>timer: {ms(this.state.time)}</h3>
-        {start}
-        {resume}
-        {stop}
-        {reset}
-      </div>
-    )
-  }
-}
-
-export default Timer;
