+++ src/components/watering.js
@@ -7,14 +7,43 @@ import Typography from '@material-ui/core/Typography';
 import PubSub from 'aws-amplify';
 import CircularProgress from '@material-ui/core/CircularProgress';
 import Grid from '@material-ui/core/Grid';
-import Timer from './timer'
+import ms from "pretty-ms"
+
 
 class Watering extends Component {
 
   constructor(props){
     super(props);
-    this.state = {load: ''}
     this.setWateringStatus = this.setWateringStatus.bind(this)
+    this.state = {
+        time: 0,
+        isOn: false,
+        start: 0,
+        load: ''
+      }
+      this.startTimer = this.startTimer.bind(this)
+      this.stopTimer = this.stopTimer.bind(this)
+      this.resetTimer = this.resetTimer.bind(this)
+  }
+
+  startTimer() {
+    this.setState({
+      isOn: true,
+      time: this.state.time,
+      start: Date.now() - this.state.time
+    })
+    this.timer = setInterval(() => this.setState({
+      time: Date.now() - this.state.start
+    }), 1);
+  }
+
+  stopTimer() {
+    this.setState({isOn: false})
+    clearInterval(this.timer)
+  }
+
+  resetTimer() {
+    this.setState({time: 0, isOn: false})
   }
 
 
@@ -43,6 +72,18 @@ class Watering extends Component {
   }
 
   setWateringStatus(precip){
+    let start = (this.state.time == 0) ?
+      <button onClick={this.startTimer}>start</button> :
+      null
+    let stop = (this.state.time == 0 || !this.state.isOn) ?
+      null :
+      <button onClick={this.stopTimer}>stop</button>
+    let resume = (this.state.time == 0 || this.state.isOn) ?
+      null :
+      <button onClick={this.startTimer}>resume</button>
+    let reset = (this.state.time == 0 || this.state.isOn) ?
+      null :
+      <button onClick={this.resetTimer}>reset</button>
     console.log(precip)
     if (precip === 0) {
       return (
@@ -69,7 +110,13 @@ class Watering extends Component {
             </Button>
           </Grid>
             <Grid item xs={1.2}>
-              <Timer />
+              <div>
+                <h3>timer: {ms(this.state.time)}</h3>
+                {start}
+                {resume}
+                {stop}
+                {reset}
+              </div>
             </Grid>
           </Grid>
       </CardContent>
